feat: add vo commands (read/write)
