+++ features/vault-operations/index.ts
@@ -2,3 +2,5 @@ export * from './vault-health.js';
 export * from './vaults-by-owner.js';
 export * from './dashboard-by-vault.js';
 export * from './vault-info.js';
+export * from './recipient.js';
+export * from './vault-overview.js';
