+++ utils/commands/index.ts
@@ -1,4 +1,2 @@
-export * from './dashboard.js';
 export * from './confirmations.js';
 export * from './report.js';
-export * from './utils.js';
