+++ utils/prompts/index.ts
@@ -2,5 +2,5 @@ export * from './default.js';
 export * from './predeposit-guarantee.js';
 export * from './vault.js';
 export * from './operations.js';
-export * from './dashboard.js';
 export * from './confirmations.js';
+export * from './mint-burn/index.js';
