+++ utils/prompts/mint-burn/confirm-mint.ts
@@ -1,0 +1,78 @@
+import { Address, formatEther } from 'viem';
+import Table from 'cli-table3';
+
+import { confirmOperation, logInfo } from 'utils';
+
+import { TABLE_PARAMS } from './constants.js';
+
+type ConfirmMintProps = {
+  vaultAddress: Address;
+  recipient: Address;
+  amountOfMint: bigint;
+  amountOfMintInStethWei: bigint;
+  newLiabilityShares: bigint;
+  newLiabilitySharesInStethWei: bigint;
+  currentLiabilityShares: bigint;
+  currentLiabilitySharesInStethWei: bigint;
+  newHealthRatio: number;
+  currentHealthRatio: number;
+  newIsHealthy: boolean;
+  currentIsHealthy: boolean;
+  type: 'shares' | 'stETH' | 'wstETH';
+};
+
+export const confirmMint = async (props: ConfirmMintProps) => {
+  const {
+    vaultAddress,
+    recipient,
+    amountOfMint,
+    amountOfMintInStethWei,
+    newLiabilityShares,
+    newLiabilitySharesInStethWei,
+    currentLiabilityShares,
+    currentLiabilitySharesInStethWei,
+    newHealthRatio,
+    currentHealthRatio,
+    newIsHealthy,
+    currentIsHealthy,
+    type,
+  } = props;
+
+  const amountOfMintInShares = formatEther(amountOfMint);
+  const amountOfMintInSteth = formatEther(amountOfMintInStethWei);
+  const isShares = type === 'shares';
+  const isWsteth = type === 'wstETH';
+
+  const value =
+    isShares || isWsteth ? amountOfMintInShares : amountOfMintInSteth;
+
+  const tableCurrent = new Table(TABLE_PARAMS);
+  logInfo(`Current vault ${vaultAddress} health:`);
+  tableCurrent.push(
+    ['Vault Address', vaultAddress],
+    ['Current Liability, wei', currentLiabilityShares],
+    ['Current Liability, Shares', formatEther(currentLiabilityShares)],
+    ['Current Liability, stETH', formatEther(currentLiabilitySharesInStethWei)],
+    ['Current Health Ratio', `${currentHealthRatio.toFixed(2)}%`],
+    ['Current Is Healthy', currentIsHealthy],
+  );
+  console.info(tableCurrent.toString());
+
+  const tableNew = new Table(TABLE_PARAMS);
+  logInfo(`Minting ${value} ${type} to ${recipient}:`);
+  tableNew.push(
+    ['Vault Address', vaultAddress],
+    ['New Liability, wei', newLiabilityShares],
+    ['New Liability, Shares', formatEther(newLiabilityShares)],
+    ['New Liability, stETH', formatEther(newLiabilitySharesInStethWei)],
+    ['New Health Ratio', `${newHealthRatio.toFixed(2)}%`],
+    ['New Is Healthy', newIsHealthy],
+  );
+  console.info(tableNew.toString());
+
+  const confirm = await confirmOperation(
+    `Are you sure you want to mint ${value} ${type} to ${recipient}?`,
+  );
+
+  return confirm;
+};
