+++ utils/prompts/mint-burn/index.ts
@@ -1,0 +1,2 @@
+export * from './confirm-mint.js';
+export * from './confirm-burn.js';
